MO MONEY DESCRIPTION
Mo-Money is a new investment firm in Canada, and it is primarily interested in deploying a new service to Canadians known as a Robo-advisor. Robo-advisors are web-based platforms that automate entire financial planning and investment services from asset allocation, purchasing investment products, rebalancing, and tax reporting with virtually no human involvement. Mo Money only deals with one specific investment product, exchange-traded funds (ETFs). These are low-cost investment products often identified by their ticker symbols on the stock exchange. For example, the Vanguard Growth Portfolio ETF is found by VGRO, and ZAG finds BMO Aggregate Bond Index ETF on the stock exchange. One important part of the Robo-advisors service is generating documents for tax reporting to the Canadian Revenue Agency (CRA). You will implement one particular component of this service known as adjusted cost base (ACB) reporting for ETFs, which is essential when filing taxes that report capital gains or losses on ETF investments. The reason for its importance is that incorrectly computing ACB may result in the client either overpaying in taxes or underpaying, both of which are undesirable for the client.